added component.mk to be compat with esp32 make system fixed a compile bug with atca_helpers.c

+++ lib/basic/atca_helpers.c
@@ -86,7 +86,7 @@ static void hex_to_lowercase(char *buffer, size_t length)
     {
         for (size_t index = 0; index < length; index++)
         {
-            buffer[index] = tolower(buffer[index]);
+            buffer[index] = tolower((int)buffer[index]);
         }
     }
 }
@@ -98,7 +98,7 @@ static void hex_to_uppercase(char *buffer, size_t length)
     {
         for (size_t index = 0; index < length; index++)
         {
-            buffer[index] = toupper(buffer[index]);
+            buffer[index] = toupper((int)buffer[index]);
         }
     }
 }
